i64 to u64, and ensuing chaos

+++ crates/sui-indexer-alt-framework/src/lib.rs
@@ -337,7 +337,7 @@ impl Indexer {
     ) -> Result<()> {
         if let (Some(watermark), Some(first_checkpoint)) = (watermark, self.first_checkpoint) {
             ensure!(
-                first_checkpoint as i64 <= watermark.checkpoint_hi_inclusive + 1,
+                first_checkpoint <= watermark.checkpoint_hi_inclusive + 1,
                 "For pipeline {}, first checkpoint override {} is too far ahead of watermark {}. \
                  This could create gaps in the data.",
                 P::NAME,
